requirements.txt
asyncpg   ~=0.24.0
ircrobots ~=0.6.0
pyyaml    ~=5.4.1

